requirements.txt
flake8==7.1.1
flake8-bugbear==24.8.19
flake8-annotations==3.1.1
flake8-docstrings==1.7.0
pep8-naming==0.14.1
flake8-builtins==2.5.0
flake8-eradicate==1.5.0
mypy==1.11.2
fastapi[all]==0.114.1
requests==2.32.3
sqlalchemy==2.0.36
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
mysql-connector-python==9.1.0